+++ cpp/libjoynr/include/joynr/Dispatcher.h
@@ -45,6 +45,7 @@ class MessagingQos;
 class RequestCaller;
 class JoynrMessage;
 class JoynrMessageSender;
+class CallContext;
 
 class JOYNR_EXPORT Dispatcher : public IDispatcher
 {