+++ cpp/tests/unit-tests/jsonserializer/JoynrJsonSerializerTest.cpp
@@ -30,6 +30,7 @@
 
 #include "joynr/serializer/Serializer.h"
 #include "joynr/Logger.h"
+#include "joynr/JoynrMessage.h"
 #include "joynr/SubscriptionPublication.h"
 #include "joynr/MulticastPublication.h"
 #include "joynr/MulticastSubscriptionRequest.h"
@@ -86,6 +87,24 @@ TEST_F(JoynrJsonSerializerTest, receivedFromGlobalAttributeIsIgnoredBySerializat
     joynr::serializer::deserializeFromJson(deserializedMsg, json);
 
     EXPECT_FALSE(deserializedMsg.isReceivedFromGlobal());
+}
+
+TEST_F(JoynrJsonSerializerTest, serializeJoynrMessage)
+{
+    // Create a JoynrMessage
+    JoynrMessage msg;
+    msg.setType("TESTTYPE");
+    std::string payload = "/67589\?\?8zhkbv\?\?\?\?\?\?\?\?\?\?L\?\?Lkj\?\?jhljvhl\?\?\?\?\?\?/\?\?\?\?\?\?";
+    msg.setPayload(payload);
+
+    // Serialize into JSON
+    std::string json = joynr::serializer::serializeToJson(msg);
+    JOYNR_LOG_TRACE(logger, "JoynrMessage JSON: {}",json);
+
+    JoynrMessage deserializedMsg;
+    joynr::serializer::deserializeFromJson(deserializedMsg, json);
+
+    EXPECT_EQ(msg, deserializedMsg);
 }
 
 TEST_F(JoynrJsonSerializerTest, exampleDeserializerApplicationException)