+++ wiki/ReleaseNotes.md
@@ -1,7 +1,9 @@
 #joynr 0.22.0
 
 ##API relevant changes
-None.
+* **[C++]** During a provider call a call context can be queried which provides the creator user id
+  field from the joynr message. Please delete the broadcastsubscriptionrequest-persistence-file and
+  subscriptionrequest-persistence-file because the file format changed.
 
 ##Other changes
 * **[JS]** Introduced mqtt messaging layer, allowing javascript runtimes including